// frontend/src/types/website.ts
export interface Testimonial {
    id: string;
    clientName: string;
    clientTitle: string;
    location?: string;
    clientImageUrl: string;
    reviewText: string;
    rating: 1 | 2 | 3 | 4 | 5;
    showOnHomepage?: boolean;
    order: number;
    createdAt?: any;
}